+++ src/main/java/roomescape/auth/application/AuthService.java
@@ -1,4 +1,4 @@
-package roomescape.auth.application.service;
+package roomescape.auth.application;
 
 import static roomescape.global.exception.ErrorMessage.INVALID_LOGIN_CREDENTIALS;
 
@@ -6,7 +6,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import roomescape.auth.application.LoginMember;
 import roomescape.auth.exception.InvalidEmailException;
 import roomescape.auth.exception.InvalidPasswordException;
 import roomescape.auth.infrastructure.TokenProvider;
