+++ src/main/java/roomescape/auth/config/AuthenticationPrincipalArgumentResolver.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
+import roomescape.auth.application.AuthService;
 import roomescape.auth.application.LoginMember;
-import roomescape.auth.application.service.AuthService;
 
 @Component
 @RequiredArgsConstructor
