+++ src/main/java/roomescape/auth/presentation/AuthController.java
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import roomescape.auth.application.AuthService;
 import roomescape.auth.application.LoginMember;
-import roomescape.auth.application.service.AuthService;
 import roomescape.auth.config.AuthenticationPrincipal;
 import roomescape.auth.presentation.dto.LoginCheckResponse;
 import roomescape.auth.presentation.dto.LoginRequest;
